requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
lxml==4.9.2
numpy==1.21.6
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pandas==1.3.5
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==20.2
pytz==2023.3
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
urllib3==1.26.15
webdriver-manager==3.8.5
webencodings==0.5.1
wsproto==1.2.0
XlsxWriter==3.0.9
